+++ noodles-bcf/examples/bcf_view_async.rs
@@ -17,12 +17,12 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
 
     let raw_header = reader.read_header().await?;
     let header = raw_header.parse()?;
-    let string_map = raw_header.parse()?;
+    let string_maps = raw_header.parse()?;
 
     let mut records = reader.records();
 
     while let Some(record) = records.try_next().await? {
-        let vcf_record = record.try_into_vcf_record(&header, &string_map)?;
+        let vcf_record = record.try_into_vcf_record(&header, &string_maps)?;
         println!("{}", vcf_record);
     }