+++ noodles-bcf/src/header.rs
@@ -1,5 +1,11 @@
 //! BCF header.
 
-mod string_map;
+pub mod string_maps;
 
-pub use self::string_map::StringMap;
+pub use self::string_maps::StringMaps;
+
+#[deprecated(
+    since = "0.11.0",
+    note = "Use `noodles_bcf::header::string_maps::StringMap` instead."
+)]
+pub use self::string_maps::StringMap;